style: change the file hierarchy, remove unnecessary types of places
