+++ src/main/kotlin/com/example/goodtripserver/places/model/PlacesResponse.kt
@@ -1,18 +1,18 @@
-package com.example.goodtripserver.places
+package com.example.goodtripserver.places.model
 
 import jakarta.validation.constraints.NotBlank
 
 data class PlacesResponse(
-    @NotBlank(message = "field \"name\" can't be null")
 
+    @NotBlank(message = "field \"name\" can't be null")
     val name: String,
 
-    val lat: Double,//мб вообще строка
+    val lat: Double,
 
     val lng: Double,
 
     val icon: String,//TODO поменять на фото
 
-    val rating: Int,
+    val rating: Int
 )
